KDoc & GET API default test
